Change to tappable scopes
