+++ app-modules/service-management/src/Models/ServiceRequest.php
@@ -22,7 +22,6 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Assist\AssistDataModel\Models\Contracts\Educatable;
 use Assist\Notifications\Models\Contracts\Subscribable;
 use Assist\AssistDataModel\Models\Contracts\Identifiable;
-use Assist\AssistDataModel\Models\Traits\EducatableScopes;
 use Illuminate\Database\UniqueConstraintViolationException;
 use Assist\Audit\Models\Concerns\Auditable as AuditableTrait;
 use Assist\Interaction\Models\Concerns\HasManyMorphedInteractions;
@@ -44,7 +43,6 @@ class ServiceRequest extends BaseModel implements Auditable, CanTriggerAutoSubsc
     use AuditableTrait;
     use HasUuids;
     use HasManyMorphedInteractions;
-    use EducatableScopes;
 
     protected $fillable = [
         'respondent_type',
