+++ app-modules/task/src/Models/Task.php
@@ -21,7 +21,6 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Assist\AssistDataModel\Models\Contracts\Educatable;
 use Assist\Notifications\Models\Contracts\Subscribable;
-use Assist\AssistDataModel\Models\Traits\EducatableScopes;
 use Assist\Audit\Models\Concerns\Auditable as AuditableTrait;
 use Assist\Campaign\Models\Contracts\ExecutableFromACampaignAction;
 use Assist\Notifications\Models\Contracts\CanTriggerAutoSubscription;
@@ -38,7 +37,6 @@ class Task extends BaseModel implements Auditable, CanTriggerAutoSubscription, E
     use AuditableTrait;
     use SoftDeletes;
     use HasStateMachine;
-    use EducatableScopes;
 
     protected $fillable = [
         'title',
