+++ app/Filament/Widgets/MyServiceRequests.php
@@ -11,6 +11,8 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Query\JoinClause;
 use Filament\Widgets\TableWidget as BaseWidget;
 use Assist\ServiceManagement\Models\ServiceRequest;
+use Assist\AssistDataModel\Models\Scopes\EducatableSort;
+use Assist\AssistDataModel\Models\Scopes\EducatableSearch;
 use Assist\ServiceManagement\Filament\Resources\ServiceRequestResource;
 
 class MyServiceRequests extends BaseWidget
@@ -41,9 +43,8 @@ class MyServiceRequests extends BaseWidget
                 TextColumn::make('respondent.display_name')
                     ->label('Related To')
                     ->getStateUsing(fn (ServiceRequest $record) => $record->respondent->{$record->respondent::displayNameKey()})
-                    ->searchable(query: fn (Builder $query, $search) => $query->educatableSearch(relationship: 'respondent', search: $search))
-                    // TODO: Find a way to get IDE to recognize educatableSort() method
-                    ->sortable(query: fn (Builder $query, string $direction): Builder => $query->educatableSort($direction)),
+                    ->searchable(query: fn (Builder $query, $search) => $query->tap(new EducatableSearch(relationship: 'respondent', search: $search)))
+                    ->sortable(query: fn (Builder $query, string $direction): Builder => $query->tap(new EducatableSort($direction))),
                 TextColumn::make('respondent.sisid')
                     ->label('SIS ID')
                     ->searchable()
